+++ src/hdr_image.rs
@@ -721,7 +721,6 @@ mod test {
         let mut img1 = HDR::new(0, 0);
         assert!(img1.normalization(Some(&1.0)).is_err());
 
-        // MODIFICATO: Rimossi i match prolissi. Ora testiamo il risultato direttamente.
         let mut img = HDR::new(1, 4);
         let mut img1 = HDR::new(1, 4);
         let mut img2 = HDR::new(1, 4);