+++ HISTORY.md
@@ -1,5 +1,6 @@
 ## Head
 
+- Add Scene Interpreter ([PR#23](https://github.com/scaioo/RayTraceRS/pull/23)).
 - Add `load_from_ldr` to `HdrImage` to support PNG/JPEG textures via
   inverse gamma correction and inverse tone mapping
   ([PR#29](https://github.com/scaioo/RayTraceRS/pull/29)).